[AIRFLOW-5783] AIP-21 Move aws redshift into providers structure (#6539)
